requirements.txt
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
colorlog==6.10.1
drf-spectacular==0.28.0
django-redis==6.0.0
celery==5.5.3
gunicorn==23.0.0
django-environ==0.12.0
pillow
django-parler